Move notifications to its own file
